Implement RedisClient
