+++ src/Illuminate/Redis/Connections/RedisClientClusterConnection.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Illuminate\Redis\Connections;
+
+class RedisClientClusterConnection extends RedisClientConnection
+{
+    //
+}
